requirements.txt
flask
Flask-SQLAlchemy
flask-login
psycopg2-binary
python-dotenv
